requirements.txt
# requirements.txt - 精简版

# ===== Web Framework =====
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.3
pydantic==2.11.7
pydantic-settings==2.10.1

# ===== LLM & Langchain =====
langchain==0.3.27
langchain-core==0.3.83
langchain-community==0.3.27
langchain-ollama==0.3.6
langchain-chroma==0.2.6
langsmith==0.4.14
ollama==0.5.3

# ===== Vector Database =====
chromadb==1.3.7

# ===== Google Maps API =====
googlemaps==4.10.0

# ===== Utilities =====
python-dotenv==1.1.1
requests==2.32.4
pyyaml==6.0.2

# ===== HTTP Client =====
httpx==0.28.1
aiohttp==3.12.15

# ===== Optional: PDF Processing (如果需要处理文档) =====
# pdfplumber==0.11.8
# pypdfium2==5.2.0